services/browser-agent/requirements.txt
a2a-sdk[http-server]>=0.3.10
claude-agent-sdk>=0.1.6
python-dotenv>=1.1.0,<2
uvicorn>=0.34.0
